+++ packages/coding-agent/test/suite/regressions/5943-session-start-notify.test.ts
@@ -34,6 +34,7 @@ function createUiContext(
 		setTitle: () => {},
 		custom: async <T>() => undefined as T,
 		pasteToEditor: () => {},
+		setImageMarkersEnabled: () => {},
 		setEditorText: () => {},
 		getEditorText: () => "",
 		editor: async () => undefined,